dptp-cm: change credentials used in image distribution
We started seeing errors like the following in dptp-cm logs:

```
imageStreamImport did not succeed:
  reason: Unauthorized,
  message: you may not have access to the container image "registry.svc.ci.openshift.org/ocp-private/4.7-priv@sha256:c692089f63cc6b7d9b875d68183e1a9a263e0a2f233d992ebe4ca5441857ef00"
```

The original secret we used for this purpose has unknown provenance and
is not managed by any automation. We should be able to use the same
secret we use in ci-operator's temporary namespace, the use-case
(allowing cross-cluster imports from api.ci) identical.